Probe naked domain and protocols for credential
When probing for a registries credential currently only the https
prefixed domain is searched. Some credential services (e.g.
docker-credential-secretservice) do not store the protocol and only
return the credential for the naked domain. This probes the naked
domain, https and http prefixed domain to see if any are stored and
returns the first match.